requirements-test.txt
django>=1.5.1
coverage
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2

# Additional test requirements go here
linaro-django-pagination
django-resort
django-tag-parser
django-with-asserts
